Inline functions to set up properties for container
We also clarify the nomenclature of "container" for the
outside "cube" - it is not a cube, since the height is different
from the side length.

The functions that were removed only make sense in the
process of creating the container and are never reused,
so we sacrifice with longer functions to get a more cohesive 
design.