Change animation logic to baking.
This means that simulation is done before animation,
and then we jump to the final frame before deleting
the container, saving and exporting.

The container is also proper referenced to try to make
sure it is deleted before exporting the STL (but after
the main file is saved, so it can be reproduced later.)

This tries to solve a bug where sometimes the container
was not being deleted when exporting the STL file,
probably because of some race condition when Blender
processes the animations.